+++ src/reset.css
@@ -32,6 +32,7 @@ ol, ul, menu {
 /* For images to not be able to exceed their container */
 img {
   max-width: 100%;
+  margin: 0 auto;
 }
 
 /* removes spacing between cells in tables */
