+++ src/styles.css
@@ -2,10 +2,10 @@
 	font-family: Arial, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu,
 		Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
   --text-color: #eeeeee;
-	--background-color: rgba(0, 0, 0, 0.6);
+	--background-color: #252d3a;
   --dark-highlight-color: #0d0057;
   --highlight-color: #57999a;
-  --bright-color: #e7c8c1;
+  --bright-color: #dd4d4d;
 }
 
 #svelte {
@@ -14,14 +14,10 @@
 	flex-direction: column;
 }
 
-html, body {
-  height: 100%;
-}
-
 body {
   font-family: var(--font-mono);
   background-color: var(--background-color);
-  background: url("https://res.cloudinary.com/tihos/image/upload/f_auto,q_auto/v1650411071/rainbow_cz6ewu.png");
+  /* background: url("https://res.cloudinary.com/tihos/image/upload/f_auto,q_auto/v1650411071/rainbow_cz6ewu.png"); */
   background-size: cover;
   color: var(--text-color);
   font-size: 16px;
@@ -35,4 +31,54 @@ a {
   text-decoration: underline;
   font-weight: bold;
   text-shadow: 0 0 0.5rem var(--dark-highlight-color);
+}
+
+button {
+  margin: 0.5rem;
+  font-weight: bold;
+  text-align: center;
+  padding: 0.5rem 1rem;
+  display: inline-block;
+  border-radius: 1rem;
+  color: var(--text-color);
+  text-transform: uppercase;
+  border: 4px solid var(--text-color);
+  background-color: var(--bright-color);
+}
+
+.stage-wrapper {
+  width: 600px;
+  display: grid;
+  flex: 1 1 550px;
+  border: 6px var(--highlight-color) solid;
+  background-color: var(--background-color);
+  box-shadow: 0 0 10px var(--dark-highlight-color);
+  border-radius: 4rem;
+  margin: 2rem auto;
+  padding: 2rem;
+  gap: 2rem;
+}
+
+.stage-wrapper h1 {
+  text-align: center;
+  font-size: 4rem;
+}
+
+.stage-wrapper div {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+@media (max-width: 600px) {
+  .stage-wrapper {
+    width: 80%;
+    margin: 2rem 1rem;
+    padding: 1rem;
+    border-radius: 2rem;
+  }
+
+  .stage-wrapper h1 {
+    font-size: 3rem;
+  }
 }
